Prefer local install. (Closes #571)
