+++ package.json
@@ -42,6 +42,7 @@
 		"github-url-from-git": "^1.5.0",
 		"has-yarn": "^2.1.0",
 		"hosted-git-info": "^3.0.0",
+		"import-local": "^3.0.2",
 		"inquirer": "^7.0.0",
 		"is-installed-globally": "^0.3.1",
 		"is-scoped": "^2.1.0",
